refactor: updated front textures of ultimate machines refactor: use fraction codec instead of float codec

+++ src/integration/kotlin/hiiragi283/ragium/client/integration/emi/recipe/HTEmiRecipe.kt
@@ -78,7 +78,7 @@ abstract class HTEmiRecipe<RECIPE : Any>(
     }
 
     protected fun addChancedOutputs(result: HTItemResultWithChance?) {
-        addOutputs(result?.base?.let(::result)?.let { stack: EmiStack -> stack.setChance(result.chance) })
+        addOutputs(result?.base?.let(::result)?.let { stack: EmiStack -> stack.setChance(result.chance.toFloat()) })
     }
 
     protected fun addOutputs(result: HTFluidResult?) {
